Replace `singular_value.dart` with `svd.dart`
